Robust checking for Bundler project
Prior to Vim version 7.3.097, Using `:call` inside `if 0` does not see that a
function returns a `dict` and throws an error for `.` as if it were string
concatenation.

This was fixed in [revision faf4b09c396e](https://code.google.com/p/vim/source/detail?r=faf4b09c396e#).

This change works around that bug with an early `return`.